Add movement
Selecting a piece now only works on that color's turn. Clicking on the move indicator dot makes the selected piece move to the indicated tile, capturing any needed pieces along the way. If you reach the diagonal on the far side of the board, your piece promotes into a king and your turn ends.